Linux compatibility
This uses the $LINUX flag during install and when initializing the shell
to create/operate a linux-friendly dotfiles environment. This assumes
that the appropriate developer tools are available (for ctags) and
that sudo is available for installing debian packages.

This updates vim settings to work with the default system-installed vim,
disabling some plugins when needed. Most should work, though.

The shell is changed to zsh by default, but both bash and zsh now default to
using starship for their prompt.