Fix support for deleting files in interactive gitpatch
When the working copy deletes a file, `git add -A` records the deletion
in the index regardless of `--intent-to-add`, which effectively stages
the deletion in git without asking on subsequent `git add -p`. What we
want is to only record new files with `--intent-to-add`, thus changing
from `-A` to `--ignore-removal` and specifying path `.`.

Since we now allow to potentially remove all files from the `right`
side, the later invocation of `git restore .` may fail with:

    error: pathspec '.' did not match any file(s) known to git

Instead, we employ `git reset --hard` to clean everything up in one go.